source-InfoBel_IT.module
this module is to search on the paginebianche.il, site to search for Italian numbers.
you can add it in the next release?
P.S.
I have a problem with this research, when it does not get results, capture Kapitol sa (which is in the bottom of the page) you can improve it? (I'm not very capable)
thanks for your work.